// src/app/pages/catalogos/interfaces/ciclos.interface.ts
export interface Ciclo {
    cveCiclo: string;
    intAnio: string;
    chrNombre: string;
    chrDescripcion: string;
    mesini: string;
    mesfin: string;
    corteIniMes1: string;
    corteFinMes1: string;
    corteMes1: string;
    corteIniMes2: string;
    corteFinMes2: string;
    corteMes2: string;
    corteIniMes3: string;
    corteFinMes3: string;
    corteMes3: string;
    diaIniExtra: string;
    diaFinExtra: string;
    mesExtra: string;
    diaIniEspe: string;
    diaFinEspe: string;
    mesEspe: string;
    fechaini: string;
    fechafin: string;
    chrStatus: string;
    
}